Add files via upload
I discover some faults at the precision flag handle within the string (%s) format specifier.  Still working on, but the main cause its because when the input calls for example "%.s, "hello"; the precision must be set on 0, and the behaviour should be nothing be printed, if the flag is "." its not printed it should be -1(non-present), but when my format parser detects . iterate one time, do not find a number to set, exit the function and the flags remains -1 (non-present).